+++ react-app/src/components/Dashboard/Feed/FeedPost/CommentForm.js
@@ -1,7 +1,7 @@
 import {useDispatch, useSelector} from "react-redux"
 import { useState } from "react";
 import "./CommentForm.css"
-import { createComment } from "../../../../store/comment";
+import { createComment } from "../../../../store/dashboard";
 
 function CommentForm({post}){
     const dispatch = useDispatch();
@@ -17,7 +17,6 @@ function CommentForm({post}){
         e.preventDefault();
         const postId = post.id
 
-        console.log("hi", comment)
 
         const payload = {
             user_id: currentUser.id,
@@ -25,6 +24,7 @@ function CommentForm({post}){
             comment,
         }
 
+    
         let newComment = await dispatch(createComment(payload))
         // .catch(async(res)=>{
         //     const data = await res.json();
